Restructured the calls to one API.
